+++ enaml/core/declarative_meta.py
@@ -5,7 +5,12 @@
 #
 # The full license is in the file LICENSE, distributed with this software.
 #------------------------------------------------------------------------------
-from atom.api import Atom, AtomMeta, DefaultValue, Member, Typed
+from atom.api import Atom, AtomMeta, ChangeType, DefaultValue, Member, Typed
+
+# The declarative engine only updates for these types of changes
+D_CHANGE_TYPES = (
+    ChangeType.UPDATE | ChangeType.PROPERTY | ChangeType.EVENT | ChangeType.DELETE
+)
 
 
 class DeclarativeDefaultHandler(Atom):
@@ -60,12 +65,11 @@ def declarative_change_handler(change):
 
     """
     # TODO think about whether this is the right place to filter on change_t
-    change_t = change['type']
-    if change_t == 'update' or change_t == 'event' or change_t == 'property':
-        owner = change['object']
-        engine = owner._d_engine
-        if engine is not None:
-            engine.write(owner, change['name'], change)
+    # NOTE: Filtering on change['type'] is done by atom
+    owner = change['object']
+    engine = owner._d_engine
+    if engine is not None:
+        engine.write(owner, change['name'], change)
 
 
 def patch_d_member(member):
@@ -88,7 +92,7 @@ def patch_d_member(member):
         new_mode = DefaultValue.CallObject_ObjectName
         member.set_default_value_mode(new_mode, handler)
     if metadata['d_readable']:
-        member.add_static_observer(declarative_change_handler)
+        member.add_static_observer(declarative_change_handler, D_CHANGE_TYPES)
 
 
 class DeclarativeMeta(AtomMeta):