+++ enaml/widgets/group_box.py
@@ -5,9 +5,9 @@
 #
 # The full license is in the file LICENSE, distributed with this software.
 #------------------------------------------------------------------------------
-from atom.api import Bool, Str, Enum, Typed, ForwardTyped, observe
+from atom.api import Bool, Str, Enum, Typed, ForwardTyped
 
-from enaml.core.declarative import d_
+from enaml.core.declarative import d_, observe
 
 from .container import Container, ProxyContainer