+++ enaml/widgets/object_combo.py
@@ -6,10 +6,10 @@
 # The full license is in the file LICENSE, distributed with this software.
 #------------------------------------------------------------------------------
 from atom.api import (
-    Bool, Callable, List, Value, Typed, ForwardTyped, set_default, observe
+    Bool, Callable, List, Value, Typed, ForwardTyped, set_default
 )
 
-from enaml.core.declarative import d_
+from enaml.core.declarative import d_, observe
 from .control import Control, ProxyControl