+++ enaml/widgets/tool_bar.py
@@ -5,9 +5,9 @@
 #
 # The full license is in the file LICENSE, distributed with this software.
 #------------------------------------------------------------------------------
-from atom.api import Bool, Enum, List, Typed, ForwardTyped, observe
+from atom.api import Bool, Enum, List, Typed, ForwardTyped
 
-from enaml.core.declarative import d_
+from enaml.core.declarative import d_, observe
 
 from .action import Action
 from .action_group import ActionGroup